refactor(fs): consolidate fileTreeUtils implementations and improve compatibility
- Merge fileTreeUtils-new.ts into fileTreeUtils.ts
- Add backward compatibility for legacy settings
- Improve error handling and logging
